+++ pom.xml
@@ -87,7 +87,6 @@
       <artifactId>async-http-client</artifactId>
       <version>2.0.34</version>
     </dependency>
-<<<<<<< HEAD
 
     <!-- test dependencies -->
     <dependency>
@@ -112,7 +111,7 @@
       <groupId>com.github.tomakehurst</groupId>
       <artifactId>wiremock</artifactId>
       <version>2.8.0</version>
-=======
+    </dependency>
     <dependency>
       <groupId>org.mockito</groupId>
       <artifactId>mockito-core</artifactId>
@@ -141,7 +140,6 @@
       <groupId>org.powermock</groupId>
       <artifactId>powermock-module-junit4-rule</artifactId>
       <version>1.7.1</version>
->>>>>>> AdvisorGlobalConfigTest part 1
       <scope>test</scope>
     </dependency>
   </dependencies>
